third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-block-indent-expected.txt

This test verifies applied indentation whenever you hit enter in "{|}" or type in "}" while inside opened block.

{} {}
  {}
  {
    
    {
        


Running: testSimpleCollapsedBlockExpanding
{
..|
}.{}
..{}
..{
....
....{
........


Running: testMulticursorCollapsedBlockExpanding
{
..|
}.{
..|
}
..{
....|
..}
..{
....
....{
........


Running: testMulticursorCollapsedBlockNotExpanding
{
|}.{}
..
..|{}
..{
....
....{
........


Running: testSingleCursorClosingBracketIndent
{}.{}
..{}
..{
..}|....
....{
........


Running: testMulticursorClosingBracketIndent
{}.{}
..{}
..{
..}|
....{
....}|


Running: testMulticursorClosingBracketIndentNotExecuted
{}.{}
..{}
..{
....}|
....{}|
........
